Делает модуляцию
