+++ index.html
@@ -8,8 +8,6 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кекстаграм</title>
-  <script src="js/functions.js" defer></script>
-  <script src="js/main.js" defer></script>
 </head>
 
 <body>
@@ -236,6 +234,9 @@
       <h2 class="data-error__title">Не удалось загрузить данные</h2>
     </section>
   </template>
+
+  <script src="js/functions.js"></script>
+  <script src="js/main.js" type="module"></script>
 
 </body>
